remove multiple node name clones in cluster 10-40% performance improvement

+++ src/core/configs/node.rs
@@ -227,7 +227,7 @@ pub(crate) enum BackendType {
 pub struct NodeConfig {
     pub(crate) log_config: Option<String>,
     pub(crate) name: Option<String>,
-    pub(crate) quorum: Option<u8>,
+    pub(crate) quorum: Option<usize>,
     pub(crate) operation_timeout: Option<String>,
     pub(crate) check_interval: Option<String>,
     pub(crate) cluster_policy: Option<String>,
@@ -488,7 +488,7 @@ impl NodeConfigYaml {
 
 pub(crate) mod tests {
     use super::*;
-    pub(crate) fn node_config(name: &str, quorum: u8) -> NodeConfig {
+    pub(crate) fn node_config(name: &str, quorum: usize) -> NodeConfig {
         NodeConfig {
             log_config: Some("".to_string()),
             name: Some(name.to_string()),
